JP Onboarding: Move Summary Steps Component Files One Level Up
